+++ packages/polaris-viz-native/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Fixed
+
+- Pass `theme` prop down to `<ChartContainer />` so `<SparkLineChart />` uses the correct Theme.
 
 ## [9.10.0] - 2023-08-03